requirements.txt
aiofile==3.5.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Brotli==1.0.9
caio==0.8.4
chardet==4.0.0
discord-py-slash-command==2.3.2
discord.py==1.7.3
expiringdict==1.2.1
idna==3.2
idna-ssl==1.1.0
motor==2.4.0
multidict==5.1.0
pymongo==3.12.0
typing-extensions==3.10.0.0
yarl==1.6.3
